+++ src/main/java/com/imap/bpm/infrastructure/repository/SequenceflowRepository.java
@@ -28,5 +28,5 @@ public interface SequenceflowRepository extends JpaRepository<Sequenceflow, UUID
     /** F4-mgmt update: borra los sequenceflows de la version antes de recrear su shape (bulk inmediato). */
     @org.springframework.data.jpa.repository.Modifying
     @org.springframework.data.jpa.repository.Query("delete from Sequenceflow s where s.processversionId = ?1")
-    long deleteByProcessversionId(UUID processversionId);
+    int deleteByProcessversionId(UUID processversionId);
 }
